fix: 追加モーダルの関数をconに移動
